Fix routes, controller_extensions, CRUDify controller, views, tests, fix footer_helper
