--- app/views/authors/email_request.html.erb
@@ -1,18 +1,0 @@
-<% @title = :author_request_title.t(title: @object.format_name) %>
-
-<%= form_tag(add_query_param(action: :email_request, id: @object.id,
-                              type: @object.type_tag)) do %>
-  <%= :author_request_note.tp %>
-
-  <div class="form-group mt-3">
-    <%= label_tag(:email_subject, :request_subject.t + ":") %>
-    <%= text_field(:email, :subject, data: {autofocus: true}, class: "form-control") %>
-  </div>
-
-  <div class="form-group mt-3">
-    <%= label_tag(:email_content, :request_message.t + ":") %>
-    <%= text_area(:email, :content, rows: 10, class: "form-control") %>
-  </div>
-
-  <%= submit_tag(:SEND.l, class: "btn btn-default") %></div><!-- .text-center --><br/>
-<% end %>
