+++ app/views/authors/reviews/show.html.erb
@@ -21,10 +21,9 @@
   <%= :review_authors_authors.t %><br/>
   <% @authors.each do |u| %>
     <%= user_link(u) %>
-    | <%= link_to(:review_authors_remove_author.t,
-                  { action: :review, id: @object.id,
-                    type: type, remove: u.id},
-                  { data: { confirm: :are_you_sure.l } }) %>
+    | <%= destroy_button(name: :review_authors_remove_author.t,
+                         target: authors_review_path(id: @object.id,
+                                   type: type, remove: u.id)) %>
     <br/>
   <% end %>
 </p>
@@ -34,8 +33,9 @@
   <% @users.each do |u|
     unless @authors.member?(u) %>
       <%= user_link(u) %>
-      | <%= link_to(:review_authors_add_author.t, action: :review,
-                    id: @object.id, type: type, add: u.id) %>
+      | <%= post_button(name: :review_authors_add_author.t,
+                        path: authors_review_path(id: @object.id,
+                                type: type, add: u.id)) %>
       <br/>
     <% end %>
   <% end %>
